+++ ChangeLog.markdown
@@ -1,5 +1,12 @@
 # Yet Another DataTables Column Filter (Yadcf) Change-log
 
+## 0.6.3
+
+* Added new filter type: text input (a simple input text) filter_type: "text" (https://github.com/vedmack/yadcf/issues/34)
+* Added new feature: case_insensitive, can be set to false or true (default) (https://github.com/vedmack/yadcf/issues/35)
+* Added new mode "startsWith" for filter_match_mode (https://github.com/vedmack/yadcf/issues/37)
+
+
 ## 0.6.0
 
 * Added integration with Chosen plugin (for the select filter)